Update i.image.bathymetry.py
Line 237 - Add potentially missing  R libraries sp, stars, and  rgeos . stars may be installed with sp, so adding may be redundant
Line 251 force use_sp, otherwise call to R raster command will fail
Line 301 make the filename variable predict a static filename created locally to the directory the script is launched from. Requires write access by user for directory launched from. This change was made to make the filename lack directory names and separators and therefore OS independent. It may be the desired behavior to write the file to scratch directory that the user is known to have write access in a manner that is os-independent.  A random integer may need to be added to the filename to prevent filename collisions.